# Optimize the execution of RuntimeProgram by saving the bool whether the instruction is feed/fetch op