每个Instruction由一个op来创建，这个op是否是feed/fetch op，其实在该Instruction定义的时候就已经确定了，不需要在`RuntimeProgram::Run()`中反复地判断。该判断在FeedGR单独RNN预测时，占了不小的一部分时间。这个从vtune的hotspots数据可以看出来。

![图片](https://agroup-bos.cdn.bcebos.com/9997a368d3cdae846708f52ef71a98e25b79e568)

这个PR在Instruction中增加一个`bool is_feed_fetch_op_`的成员变量和一个`bool is_feed_fetch_op() const`的成员函数，来避免反复的`std::string::compare`操作。

- 基于develop测试FeedGR单步RNN的性能：

| 优化前 | 优化后 |
|---|---|
| 0.0423 ms | 0.0396 ms |

- 基于https://github.com/PaddlePaddle/Paddle-Lite/pull/2679 测试FeedGR单步RNN的性能：

| #2679 | #2679 + #2703 |
|---|---|
| 0.0255 ms | 0.0216 ms |